+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Fixed `source` view supported syntaxes to list modes with no mime
+- Fixed editor's hint popup positioning when widget is not fit into the page bounds
 
 ## 1.0.0-beta.57 (04-03-2021)